feat: updated process uptime and other minor enhancements

+++ include/processor.h
@@ -4,11 +4,11 @@
 
 class Processor {
  public:
-  float Utilization();
+  double Utilization();
 
  private:
-  float cpu_idle[2]={0.0,0.0};
-  float cpu_nonidle[2]={0.0,0.0};
+  double cpu_idle[2]={0.0,0.0};
+  double cpu_nonidle[2]={0.0,0.0};
 };
 
 #endif
